Refactor quiz engine into dedicated modules
